\name{SKATbe-package}
\alias{SKATbe-package}
\alias{SKATbe}
\docType{package}
\title{
Skat and burden tests with backwards elimination.
}
\description{
Skat and burden tests are done with backwards elimination to select a set of likely disease related variants. Subsampling is used to speed computation when using datasets with a large number of variants. The main functions are skat.be(), which is intended for use within an interactive R session, and skat.cl(), which takes file names as arguments and is designed for convenient use in command line scripts.
}
\details{
\tabular{ll}{
Package: \tab SKAT_BE_package\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-31\cr
License: \tab What license is it under?\cr
}
}
\author{
Iuliana Ionita-Laza, Kenneth J. McCallum

Maintainer: Iuliana Ionita-Laza <ii2135@columbia.edu>

}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
